dev-requirements.txt
flake8==4.0.1
pytest-django==4.5.0
pytest-factoryboy==2.3.0
coverage==6.4.0
django-debug-toolbar==3.4
django-import-export==2.3.0
